requirements.txt
beautifulsoup4
fastapi
numpy
pandas
psycopg2-binary
pydantic
pymupdf
python-dotenv
qdrant-client
rank-bm25
requests
sentence-transformers
sqlalchemy
uvicorn
python-multipart

# LangChain pieces used by ingestion/chunking
langchain-core
langchain-text-splitters

# LLM client used in generation/HyDE
groq
